+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 - ``NeuronCriteria``: Added ``cellBodyFiber`` parameter.
 - ``SynapseCriteria``: Changed the default value of ``primary_only`` to ``False``,
   since it may been counter-intuitive to obtain duplicate results by default.
+- Added mitochondria queries
 
 0.4.12 / 2020-11-21
 -------------------